Refactor parser tests into separate file
Move all tests from parser.rs (2504 lines) into a dedicated tests.rs module:

- src/parser.rs: Reduced from 2504 to 107 lines
- src/parser/tests.rs: All parser tests (~2400 lines)

This refactoring improves code organization by:
- Separating test code from implementation
- Making the main parser module much more readable
- Following Rust module conventions for test organization
- Maintaining all existing functionality and test coverage
